+++ tests/components/automation/test_recorder.py
@@ -11,7 +11,7 @@ from homeassistant.components.automation import (
     ATTR_MODE,
     CONF_ID,
 )
-from homeassistant.components.recorder.models import StateAttributes, States
+from homeassistant.components.recorder.db_schema import StateAttributes, States
 from homeassistant.components.recorder.util import session_scope
 from homeassistant.const import ATTR_ENTITY_ID, ATTR_FRIENDLY_NAME
 from homeassistant.core import State
