+++ tests/components/climate/test_recorder.py
@@ -15,7 +15,7 @@ from homeassistant.components.climate.const import (
     ATTR_SWING_MODES,
     ATTR_TARGET_TEMP_STEP,
 )
-from homeassistant.components.recorder.models import StateAttributes, States
+from homeassistant.components.recorder.db_schema import StateAttributes, States
 from homeassistant.components.recorder.util import session_scope
 from homeassistant.const import ATTR_FRIENDLY_NAME
 from homeassistant.core import State
