+++ tests/components/input_text/test_recorder.py
@@ -11,7 +11,7 @@ from homeassistant.components.input_text import (
     DOMAIN,
     MODE_TEXT,
 )
-from homeassistant.components.recorder.models import StateAttributes, States
+from homeassistant.components.recorder.db_schema import StateAttributes, States
 from homeassistant.components.recorder.util import session_scope
 from homeassistant.const import ATTR_EDITABLE
 from homeassistant.core import HomeAssistant, State
