+++ tests/components/recorder/common.py
@@ -14,13 +14,13 @@ from homeassistant import core as ha
 from homeassistant.components import recorder
 from homeassistant.components.recorder import get_instance, statistics
 from homeassistant.components.recorder.core import Recorder
-from homeassistant.components.recorder.models import RecorderRuns
+from homeassistant.components.recorder.db_schema import RecorderRuns
 from homeassistant.components.recorder.tasks import RecorderTask, StatisticsTask
 from homeassistant.core import HomeAssistant
 from homeassistant.util import dt as dt_util
 
 from tests.common import async_fire_time_changed, fire_time_changed
-from tests.components.recorder import models_schema_0
+from tests.components.recorder import db_schema_0
 
 DEFAULT_PURGE_TASKS = 3
 
@@ -122,7 +122,7 @@ def create_engine_test(*args, **kwargs):
     This simulates an existing db with the old schema.
     """
     engine = create_engine(*args, **kwargs)
-    models_schema_0.Base.metadata.create_all(engine)
+    db_schema_0.Base.metadata.create_all(engine)
     return engine
