+++ tests/components/recorder/db_schema_0.py
