+++ tests/components/recorder/db_schema_16.py
