+++ tests/components/recorder/db_schema_18.py
