+++ tests/components/recorder/db_schema_22.py
