+++ tests/components/recorder/db_schema_23.py
