+++ tests/components/recorder/db_schema_23_with_newer_columns.py
