+++ tests/components/recorder/db_schema_28.py
