+++ tests/components/recorder/test_statistics.py
@@ -13,10 +13,8 @@ from sqlalchemy.orm import Session
 from homeassistant.components import recorder
 from homeassistant.components.recorder import history, statistics
 from homeassistant.components.recorder.const import DATA_INSTANCE, SQLITE_URL_PREFIX
-from homeassistant.components.recorder.models import (
-    StatisticsShortTerm,
-    process_timestamp_to_utc_isoformat,
-)
+from homeassistant.components.recorder.db_schema import StatisticsShortTerm
+from homeassistant.components.recorder.models import process_timestamp_to_utc_isoformat
 from homeassistant.components.recorder.statistics import (
     async_add_external_statistics,
     delete_statistics_duplicates,
@@ -390,7 +388,7 @@ def test_rename_entity_collision(hass_recorder, caplog):
     }
 
     with session_scope(hass=hass) as session:
-        session.add(recorder.models.StatisticsMeta.from_meta(metadata_1))
+        session.add(recorder.db_schema.StatisticsMeta.from_meta(metadata_1))
 
     # Rename entity sensor.test1 to sensor.test99
     @callback
@@ -941,7 +939,7 @@ def test_duplicate_statistics_handle_integrity_error(hass_recorder, caplog):
         assert insert_statistics_mock.call_count == 3
 
     with session_scope(hass=hass) as session:
-        tmp = session.query(recorder.models.Statistics).all()
+        tmp = session.query(recorder.db_schema.Statistics).all()
         assert len(tmp) == 2
 
     assert "Blocked attempt to insert duplicated statistic rows" in caplog.text
@@ -952,15 +950,19 @@ def _create_engine_28(*args, **kwargs):
 
     This simulates an existing db with the old schema.
     """
-    module = "tests.components.recorder.models_schema_28"
+    module = "tests.components.recorder.db_schema_28"
     importlib.import_module(module)
-    old_models = sys.modules[module]
+    old_db_schema = sys.modules[module]
     engine = create_engine(*args, **kwargs)
-    old_models.Base.metadata.create_all(engine)
+    old_db_schema.Base.metadata.create_all(engine)
     with Session(engine) as session:
-        session.add(recorder.models.StatisticsRuns(start=statistics.get_start_time()))
         session.add(
-            recorder.models.SchemaChanges(schema_version=old_models.SCHEMA_VERSION)
+            recorder.db_schema.StatisticsRuns(start=statistics.get_start_time())
+        )
+        session.add(
+            recorder.db_schema.SchemaChanges(
+                schema_version=old_db_schema.SCHEMA_VERSION
+            )
         )
         session.commit()
     return engine
@@ -971,9 +973,9 @@ def test_delete_metadata_duplicates(caplog, tmpdir):
     test_db_file = tmpdir.mkdir("sqlite").join("test_run_info.db")
     dburl = f"{SQLITE_URL_PREFIX}//{test_db_file}"
 
-    module = "tests.components.recorder.models_schema_28"
+    module = "tests.components.recorder.db_schema_28"
     importlib.import_module(module)
-    old_models = sys.modules[module]
+    old_db_schema = sys.modules[module]
 
     external_energy_metadata_1 = {
         "has_mean": False,
@@ -1001,8 +1003,8 @@ def test_delete_metadata_duplicates(caplog, tmpdir):
     }
 
     # Create some duplicated statistics_meta with schema version 28
-    with patch.object(recorder, "models", old_models), patch.object(
-        recorder.migration, "SCHEMA_VERSION", old_models.SCHEMA_VERSION
+    with patch.object(recorder, "db_schema", old_db_schema), patch.object(
+        recorder.migration, "SCHEMA_VERSION", old_db_schema.SCHEMA_VERSION
     ), patch(
         "homeassistant.components.recorder.core.create_engine", new=_create_engine_28
     ):
@@ -1013,15 +1015,17 @@ def test_delete_metadata_duplicates(caplog, tmpdir):
 
         with session_scope(hass=hass) as session:
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_1)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_1)
             )
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_2)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_2)
             )
-            session.add(recorder.models.StatisticsMeta.from_meta(external_co2_metadata))
+            session.add(
+                recorder.db_schema.StatisticsMeta.from_meta(external_co2_metadata)
+            )
 
         with session_scope(hass=hass) as session:
-            tmp = session.query(recorder.models.StatisticsMeta).all()
+            tmp = session.query(recorder.db_schema.StatisticsMeta).all()
             assert len(tmp) == 3
             assert tmp[0].id == 1
             assert tmp[0].statistic_id == "test:total_energy_import_tariff_1"
@@ -1042,7 +1046,7 @@ def test_delete_metadata_duplicates(caplog, tmpdir):
 
     assert "Deleted 1 duplicated statistics_meta rows" in caplog.text
     with session_scope(hass=hass) as session:
-        tmp = session.query(recorder.models.StatisticsMeta).all()
+        tmp = session.query(recorder.db_schema.StatisticsMeta).all()
         assert len(tmp) == 2
         assert tmp[0].id == 2
         assert tmp[0].statistic_id == "test:total_energy_import_tariff_1"
@@ -1058,9 +1062,9 @@ def test_delete_metadata_duplicates_many(caplog, tmpdir):
     test_db_file = tmpdir.mkdir("sqlite").join("test_run_info.db")
     dburl = f"{SQLITE_URL_PREFIX}//{test_db_file}"
 
-    module = "tests.components.recorder.models_schema_28"
+    module = "tests.components.recorder.db_schema_28"
     importlib.import_module(module)
-    old_models = sys.modules[module]
+    old_db_schema = sys.modules[module]
 
     external_energy_metadata_1 = {
         "has_mean": False,
@@ -1088,8 +1092,8 @@ def test_delete_metadata_duplicates_many(caplog, tmpdir):
     }
 
     # Create some duplicated statistics with schema version 28
-    with patch.object(recorder, "models", old_models), patch.object(
-        recorder.migration, "SCHEMA_VERSION", old_models.SCHEMA_VERSION
+    with patch.object(recorder, "db_schema", old_db_schema), patch.object(
+        recorder.migration, "SCHEMA_VERSION", old_db_schema.SCHEMA_VERSION
     ), patch(
         "homeassistant.components.recorder.core.create_engine", new=_create_engine_28
     ):
@@ -1100,20 +1104,26 @@ def test_delete_metadata_duplicates_many(caplog, tmpdir):
 
         with session_scope(hass=hass) as session:
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_1)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_1)
             )
             for _ in range(3000):
                 session.add(
-                    recorder.models.StatisticsMeta.from_meta(external_energy_metadata_1)
+                    recorder.db_schema.StatisticsMeta.from_meta(
+                        external_energy_metadata_1
+                    )
                 )
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_2)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_2)
             )
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_2)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_2)
             )
-            session.add(recorder.models.StatisticsMeta.from_meta(external_co2_metadata))
-            session.add(recorder.models.StatisticsMeta.from_meta(external_co2_metadata))
+            session.add(
+                recorder.db_schema.StatisticsMeta.from_meta(external_co2_metadata)
+            )
+            session.add(
+                recorder.db_schema.StatisticsMeta.from_meta(external_co2_metadata)
+            )
 
         hass.stop()
         dt_util.DEFAULT_TIME_ZONE = ORIG_TZ
@@ -1127,7 +1137,7 @@ def test_delete_metadata_duplicates_many(caplog, tmpdir):
 
     assert "Deleted 3002 duplicated statistics_meta rows" in caplog.text
     with session_scope(hass=hass) as session:
-        tmp = session.query(recorder.models.StatisticsMeta).all()
+        tmp = session.query(recorder.db_schema.StatisticsMeta).all()
         assert len(tmp) == 3
         assert tmp[0].id == 3001
         assert tmp[0].statistic_id == "test:total_energy_import_tariff_1"
