+++ tests/components/recorder/test_statistics_v23_migration.py
@@ -25,7 +25,7 @@ from tests.components.recorder.common import wait_recording_done
 ORIG_TZ = dt_util.DEFAULT_TIME_ZONE
 
 CREATE_ENGINE_TARGET = "homeassistant.components.recorder.core.create_engine"
-SCHEMA_MODULE = "tests.components.recorder.models_schema_23_with_newer_columns"
+SCHEMA_MODULE = "tests.components.recorder.db_schema_23_with_newer_columns"
 
 
 def _create_engine_test(*args, **kwargs):
@@ -34,13 +34,17 @@ def _create_engine_test(*args, **kwargs):
     This simulates an existing db with the old schema.
     """
     importlib.import_module(SCHEMA_MODULE)
-    old_models = sys.modules[SCHEMA_MODULE]
+    old_db_schema = sys.modules[SCHEMA_MODULE]
     engine = create_engine(*args, **kwargs)
-    old_models.Base.metadata.create_all(engine)
+    old_db_schema.Base.metadata.create_all(engine)
     with Session(engine) as session:
-        session.add(recorder.models.StatisticsRuns(start=statistics.get_start_time()))
         session.add(
-            recorder.models.SchemaChanges(schema_version=old_models.SCHEMA_VERSION)
+            recorder.db_schema.StatisticsRuns(start=statistics.get_start_time())
+        )
+        session.add(
+            recorder.db_schema.SchemaChanges(
+                schema_version=old_db_schema.SCHEMA_VERSION
+            )
         )
         session.commit()
     return engine
@@ -52,7 +56,7 @@ def test_delete_duplicates(caplog, tmpdir):
     dburl = f"{SQLITE_URL_PREFIX}//{test_db_file}"
 
     importlib.import_module(SCHEMA_MODULE)
-    old_models = sys.modules[SCHEMA_MODULE]
+    old_db_schema = sys.modules[SCHEMA_MODULE]
 
     period1 = dt_util.as_utc(dt_util.parse_datetime("2021-09-01 00:00:00"))
     period2 = dt_util.as_utc(dt_util.parse_datetime("2021-09-30 23:00:00"))
@@ -171,8 +175,8 @@ def test_delete_duplicates(caplog, tmpdir):
     }
 
     # Create some duplicated statistics with schema version 23
-    with patch.object(recorder, "models", old_models), patch.object(
-        recorder.migration, "SCHEMA_VERSION", old_models.SCHEMA_VERSION
+    with patch.object(recorder, "db_schema", old_db_schema), patch.object(
+        recorder.migration, "SCHEMA_VERSION", old_db_schema.SCHEMA_VERSION
     ), patch(CREATE_ENGINE_TARGET, new=_create_engine_test):
         hass = get_test_home_assistant()
         setup_component(hass, "recorder", {"recorder": {"db_url": dburl}})
@@ -181,19 +185,21 @@ def test_delete_duplicates(caplog, tmpdir):
 
         with session_scope(hass=hass) as session:
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_1)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_1)
             )
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_2)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_2)
             )
-            session.add(recorder.models.StatisticsMeta.from_meta(external_co2_metadata))
+            session.add(
+                recorder.db_schema.StatisticsMeta.from_meta(external_co2_metadata)
+            )
         with session_scope(hass=hass) as session:
             for stat in external_energy_statistics_1:
-                session.add(recorder.models.Statistics.from_stats(1, stat))
+                session.add(recorder.db_schema.Statistics.from_stats(1, stat))
             for stat in external_energy_statistics_2:
-                session.add(recorder.models.Statistics.from_stats(2, stat))
+                session.add(recorder.db_schema.Statistics.from_stats(2, stat))
             for stat in external_co2_statistics:
-                session.add(recorder.models.Statistics.from_stats(3, stat))
+                session.add(recorder.db_schema.Statistics.from_stats(3, stat))
 
         hass.stop()
         dt_util.DEFAULT_TIME_ZONE = ORIG_TZ
@@ -218,7 +224,7 @@ def test_delete_duplicates_many(caplog, tmpdir):
     dburl = f"{SQLITE_URL_PREFIX}//{test_db_file}"
 
     importlib.import_module(SCHEMA_MODULE)
-    old_models = sys.modules[SCHEMA_MODULE]
+    old_db_schema = sys.modules[SCHEMA_MODULE]
 
     period1 = dt_util.as_utc(dt_util.parse_datetime("2021-09-01 00:00:00"))
     period2 = dt_util.as_utc(dt_util.parse_datetime("2021-09-30 23:00:00"))
@@ -337,8 +343,8 @@ def test_delete_duplicates_many(caplog, tmpdir):
     }
 
     # Create some duplicated statistics with schema version 23
-    with patch.object(recorder, "models", old_models), patch.object(
-        recorder.migration, "SCHEMA_VERSION", old_models.SCHEMA_VERSION
+    with patch.object(recorder, "db_schema", old_db_schema), patch.object(
+        recorder.migration, "SCHEMA_VERSION", old_db_schema.SCHEMA_VERSION
     ), patch(CREATE_ENGINE_TARGET, new=_create_engine_test):
         hass = get_test_home_assistant()
         setup_component(hass, "recorder", {"recorder": {"db_url": dburl}})
@@ -347,25 +353,27 @@ def test_delete_duplicates_many(caplog, tmpdir):
 
         with session_scope(hass=hass) as session:
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_1)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_1)
             )
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_2)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_2)
             )
-            session.add(recorder.models.StatisticsMeta.from_meta(external_co2_metadata))
+            session.add(
+                recorder.db_schema.StatisticsMeta.from_meta(external_co2_metadata)
+            )
         with session_scope(hass=hass) as session:
             for stat in external_energy_statistics_1:
-                session.add(recorder.models.Statistics.from_stats(1, stat))
+                session.add(recorder.db_schema.Statistics.from_stats(1, stat))
             for _ in range(3000):
                 session.add(
-                    recorder.models.Statistics.from_stats(
+                    recorder.db_schema.Statistics.from_stats(
                         1, external_energy_statistics_1[-1]
                     )
                 )
             for stat in external_energy_statistics_2:
-                session.add(recorder.models.Statistics.from_stats(2, stat))
+                session.add(recorder.db_schema.Statistics.from_stats(2, stat))
             for stat in external_co2_statistics:
-                session.add(recorder.models.Statistics.from_stats(3, stat))
+                session.add(recorder.db_schema.Statistics.from_stats(3, stat))
 
         hass.stop()
         dt_util.DEFAULT_TIME_ZONE = ORIG_TZ
@@ -391,7 +399,7 @@ def test_delete_duplicates_non_identical(caplog, tmpdir):
     dburl = f"{SQLITE_URL_PREFIX}//{test_db_file}"
 
     importlib.import_module(SCHEMA_MODULE)
-    old_models = sys.modules[SCHEMA_MODULE]
+    old_db_schema = sys.modules[SCHEMA_MODULE]
 
     period1 = dt_util.as_utc(dt_util.parse_datetime("2021-09-01 00:00:00"))
     period2 = dt_util.as_utc(dt_util.parse_datetime("2021-09-30 23:00:00"))
@@ -480,8 +488,8 @@ def test_delete_duplicates_non_identical(caplog, tmpdir):
     }
 
     # Create some duplicated statistics with schema version 23
-    with patch.object(recorder, "models", old_models), patch.object(
-        recorder.migration, "SCHEMA_VERSION", old_models.SCHEMA_VERSION
+    with patch.object(recorder, "db_schema", old_db_schema), patch.object(
+        recorder.migration, "SCHEMA_VERSION", old_db_schema.SCHEMA_VERSION
     ), patch(CREATE_ENGINE_TARGET, new=_create_engine_test):
         hass = get_test_home_assistant()
         setup_component(hass, "recorder", {"recorder": {"db_url": dburl}})
@@ -490,16 +498,16 @@ def test_delete_duplicates_non_identical(caplog, tmpdir):
 
         with session_scope(hass=hass) as session:
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_1)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_1)
             )
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_2)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_2)
             )
         with session_scope(hass=hass) as session:
             for stat in external_energy_statistics_1:
-                session.add(recorder.models.Statistics.from_stats(1, stat))
+                session.add(recorder.db_schema.Statistics.from_stats(1, stat))
             for stat in external_energy_statistics_2:
-                session.add(recorder.models.Statistics.from_stats(2, stat))
+                session.add(recorder.db_schema.Statistics.from_stats(2, stat))
 
         hass.stop()
         dt_util.DEFAULT_TIME_ZONE = ORIG_TZ
@@ -560,7 +568,7 @@ def test_delete_duplicates_short_term(caplog, tmpdir):
     dburl = f"{SQLITE_URL_PREFIX}//{test_db_file}"
 
     importlib.import_module(SCHEMA_MODULE)
-    old_models = sys.modules[SCHEMA_MODULE]
+    old_db_schema = sys.modules[SCHEMA_MODULE]
 
     period4 = dt_util.as_utc(dt_util.parse_datetime("2021-10-31 23:00:00"))
 
@@ -580,8 +588,8 @@ def test_delete_duplicates_short_term(caplog, tmpdir):
     }
 
     # Create some duplicated statistics with schema version 23
-    with patch.object(recorder, "models", old_models), patch.object(
-        recorder.migration, "SCHEMA_VERSION", old_models.SCHEMA_VERSION
+    with patch.object(recorder, "db_schema", old_db_schema), patch.object(
+        recorder.migration, "SCHEMA_VERSION", old_db_schema.SCHEMA_VERSION
     ), patch(CREATE_ENGINE_TARGET, new=_create_engine_test):
         hass = get_test_home_assistant()
         setup_component(hass, "recorder", {"recorder": {"db_url": dburl}})
@@ -590,14 +598,14 @@ def test_delete_duplicates_short_term(caplog, tmpdir):
 
         with session_scope(hass=hass) as session:
             session.add(
-                recorder.models.StatisticsMeta.from_meta(external_energy_metadata_1)
+                recorder.db_schema.StatisticsMeta.from_meta(external_energy_metadata_1)
             )
         with session_scope(hass=hass) as session:
             session.add(
-                recorder.models.StatisticsShortTerm.from_stats(1, statistic_row)
+                recorder.db_schema.StatisticsShortTerm.from_stats(1, statistic_row)
             )
             session.add(
-                recorder.models.StatisticsShortTerm.from_stats(1, statistic_row)
+                recorder.db_schema.StatisticsShortTerm.from_stats(1, statistic_row)
             )
 
         hass.stop()
