+++ tests/components/script/test_recorder.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import pytest
 
 from homeassistant.components import script
-from homeassistant.components.recorder.models import StateAttributes, States
+from homeassistant.components.recorder.db_schema import StateAttributes, States
 from homeassistant.components.recorder.util import session_scope
 from homeassistant.components.script import (
     ATTR_CUR,
